Check that files are only included once, fix bug
The bug here was that $hfiles{$hfiles} was being set rather than
$hfiles{$hfile}, so the mechanism to prevent double includes was broken.

Added a test xt/c-file-check.t. This is fragile against the format of
the .h files in tidy-lib, but assuming this doesn't change it works ok
and the bug is now gone for all .h files.

New file disabled-options to list options which need to be disabled
from the interface.